milestone4-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pandas==2.0.0
scikit-learn==1.3.0
neo4j==5.13.0
sentence-transformers==2.2.2
pinecone==4.0.0
numpy==1.24.0
